reworking update methods for ratings and reviews

+++ Spring_Boot/SpringBoot/movie-rental/src/main/java/com/teksystems/bootcamp/springboot/movierental/repository/RatingRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RatingRepository extends JpaRepository<Rating, Integer> {
+public interface RatingRepository extends JpaRepository<Rating, Short> {
 }
